+++ src/main/java/com/zemiak/books/phone/domain/WebPage.java
@@ -13,7 +13,7 @@ import javax.xml.bind.annotation.XmlTransient;
 
 @XmlAccessorType(XmlAccessType.FIELD)
 @XmlRootElement
-public class WebPage implements Comparable {
+public class WebPage implements Comparable, AutoCloseable {
     private String name;
     private URL url;
     private int id;
@@ -22,24 +22,32 @@ public class WebPage implements Comparable {
     private Author author = null;
 
     @XmlTransient
-    private Client client;
+    private Client client = null;
     
     public WebPage() {
-        ClientConfig config = new DefaultClientConfig();
-        client = Client.create(config);
     }
 
     public Author getAuthor() {
         if (null == author) {
-            WebResource resource = client.resource(authorUrl);
+            WebResource resource = getClient().resource(authorUrl);
             author = resource.get(Author.class);
         }
 
         return author;
     }
 
+    private Client getClient() {
+        if (null == client) {
+            ClientConfig config = new DefaultClientConfig();
+            client = Client.create(config);
+        }
+        
+        return client;
+    }
+
+    @Override
     public void close() {
-        client.destroy();
+        if (null != client) client.destroy();
     }
 
     public String getName() {
